As per feedback, move offset tool to /london/offsets, replace API route with server action
