+++ src/app/(game)/london/page.tsx
@@ -21,7 +21,7 @@ const font = Cabin({
 export const metadata = config.METADATA
 
 export default async function London() {
-  const { routesFc, savedSettings } = loadCityRoutes('london', config)
+  const routesFc = loadCityRoutes('london')
 
   const fc = {
     ...data,
@@ -31,7 +31,7 @@ export default async function London() {
   return (
     <Provider value={config}>
       <Main className={`${font.className} min-h-screen`}>
-        <GamePage fc={fc} routes={routesFc} savedSettings={savedSettings} />
+        <GamePage fc={fc} routes={routesFc} />
       </Main>
     </Provider>
   )
