+++ src/components/GamePage.tsx
@@ -11,15 +11,8 @@ import useNormalizeString from '@/hooks/useNormalizeString'
 import useTranslation from '@/hooks/useTranslation'
 import { useConfig } from '@/lib/configContext'
 import {
-  buildSegmentedRoutes,
-  computeInterpolatedOffsets,
-  detectOverlapSections,
-  processRoutesForSameLineOverlaps
-} from '@/lib/lineOffsetting'
-import {
   DataFeature,
   DataFeatureCollection,
-  ProcessedSettings,
   RoutesFeatureCollection,
 } from '@/lib/types'
 import { useLocalStorageValue } from '@react-hookz/web'
@@ -31,18 +24,14 @@ import 'mapbox-gl/dist/mapbox-gl.css'
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import 'react-circular-progressbar/dist/styles.css'
 
-const DEFAULT_STEP_SIZE = 3
-
 export default function GamePage({
   fc,
   routes,
-  savedSettings,
 }: {
   fc: DataFeatureCollection
   routes?: RoutesFeatureCollection
-  savedSettings?: ProcessedSettings
 }) {
-  const { BEG_THRESHOLD, CITY_NAME, MAP_CONFIG, LINES, MAP_FROM_DATA, OFFSET_PROCESSING_MODE } =
+  const { BEG_THRESHOLD, CITY_NAME, MAP_CONFIG, LINES, MAP_FROM_DATA } =
     useConfig()
   const { t } = useTranslation()
 
@@ -51,33 +40,8 @@ export default function GamePage({
   const [map, setMap] = useState<mapboxgl.Map | null>(null)
   const [hoveredId, setHoveredId] = useState<number | null>(null)
   const inputRef = useRef<HTMLInputElement | null>(null)
-  const filteredRoutesRef = useRef<typeof routes | undefined>(undefined)
   const { hideLabels, setHideLabels } = useHideLabels(map)
   const [showStripeModal, setShowStripeModal] = useState<boolean>(false)
-  const [routeStates, setRouteStates] = useState<Map<string, 'hidden' | 'magenta'>>(new Map())
-  const [segmentOffsets, setSegmentOffsets] = useState<Map<string, number>>(new Map())
-  const [globalStepSize, setGlobalStepSize] = useState<number>(3)
-  const [expandedLines, setExpandedLines] = useState<Set<string>>(new Set())
-  const [saving, setSaving] = useState(false)
-  const [saveSuccess, setSaveSuccess] = useState(false)
-  // Mirrors what's actually on disk — starts from the server-loaded prop, updated after each save
-  const [latestSavedSettings, setLatestSavedSettings] = useState<ProcessedSettings | undefined>(savedSettings)
-
-  // Show overlay while map is initialising (startup + any post-save reinit) — dev tool only
-  const showSavingOverlay = !!OFFSET_PROCESSING_MODE && (!map || saving)
-
-  const hasChanges = segmentOffsets.size > 0 || globalStepSize !== DEFAULT_STEP_SIZE
-
-  const savedSettingsMatchCurrent = useMemo(() => {
-    if (!latestSavedSettings) return false
-    if (latestSavedSettings.globalStepSize !== globalStepSize) return false
-    const saved = latestSavedSettings.segmentOffsets
-    if (Object.keys(saved).length !== segmentOffsets.size) return false
-    for (const [k, v] of segmentOffsets) {
-      if (saved[k] !== v) return false
-    }
-    return true
-  }, [latestSavedSettings, segmentOffsets, globalStepSize])
 
   const { value: hasShownStripeModal, set: setHasShownStripeModal } =
     useLocalStorageValue<boolean>('has-shown-stripe-modal', {
@@ -178,106 +142,6 @@ export default function GamePage({
 
   const foundProportion = found.length / fc.features.length
 
-  // Station positions for transition splitting
-  const stationsForOffsets = useMemo(() => {
-    return fc.features
-      .filter(f => f.geometry.type === 'Point')
-      .map(f => ({ coord: f.geometry.coordinates as number[] }))
-  }, [fc])
-
-  // Step 1: Deduplicate same-line overlaps
-  const deduped = useMemo(() => {
-    if (!routes || !MAP_FROM_DATA || !OFFSET_PROCESSING_MODE) return undefined
-    return processRoutesForSameLineOverlaps(routes, LINES)
-  }, [routes, LINES, MAP_FROM_DATA, OFFSET_PROCESSING_MODE])
-
-  // Step 2: Detect overlap sections between different lines
-  const overlapSections = useMemo(() => {
-    if (!deduped) return []
-    return detectOverlapSections(deduped)
-  }, [deduped])
-
-  // Step 3: Build named segments (no auto offsets); split transition segments at stations
-  const segmentedRoutes = useMemo(() => {
-    if (!deduped || !MAP_FROM_DATA || !OFFSET_PROCESSING_MODE) return undefined
-    return buildSegmentedRoutes(deduped, overlapSections, LINES, stationsForOffsets)
-  }, [deduped, overlapSections, LINES, MAP_FROM_DATA, stationsForOffsets])
-
-  const segmentInfoMap = segmentedRoutes?.segments
-  const segmentsByLine = segmentedRoutes?.segmentsByLine
-  const overlapSegmentNames = segmentedRoutes?.overlapSegmentNames
-
-  // Step 4: Compute interpolated offsets
-  const interpolatedOffsets = useMemo(() => {
-    if (!segmentedRoutes) return new Map<string, number>()
-    return computeInterpolatedOffsets(segmentedRoutes.segmentsByLine, segmentOffsets)
-  }, [segmentedRoutes, segmentOffsets])
-
-  // Step 5: Stamp overlapOffsetPx on each feature
-  const processedRoutes = useMemo(() => {
-    if (!segmentedRoutes) return undefined
-    const features = segmentedRoutes.fc.features.map(f => {
-      const segName = (f.properties as any)?.segmentName as string
-      const unitOffset = interpolatedOffsets.get(segName) ?? 0
-      return { ...f, properties: { ...f.properties, overlapOffsetPx: unitOffset * globalStepSize } }
-    })
-    return { ...segmentedRoutes.fc, features } as typeof routes
-  }, [segmentedRoutes, interpolatedOffsets, globalStepSize])
-
-  useEffect(() => {
-    setSaveSuccess(false)
-  }, [segmentOffsets, globalStepSize])
-
-  const handleLoadSavedSettings = useCallback(() => {
-    if (!latestSavedSettings) return
-    setSegmentOffsets(new Map(Object.entries(latestSavedSettings.segmentOffsets)))
-    setGlobalStepSize(latestSavedSettings.globalStepSize)
-  }, [latestSavedSettings])
-
-  const handleReset = useCallback(() => {
-    if (!confirm('Reset all offset settings back to zero?')) return
-    setSegmentOffsets(new Map())
-    setGlobalStepSize(DEFAULT_STEP_SIZE)
-  }, [])
-
-  const handleSaveRoutes = useCallback(async () => {
-    if (!processedRoutes) return
-    setSaving(true)
-    setSaveSuccess(false)
-    const newSettings: ProcessedSettings = {
-      segmentOffsets: Object.fromEntries(segmentOffsets),
-      globalStepSize,
-    }
-    try {
-      const res = await fetch('/api/save-routes', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify({ city: CITY_NAME, data: processedRoutes, settings: newSettings }),
-      })
-      if (res.ok) {
-        setLatestSavedSettings(newSettings)
-        setSaveSuccess(true)
-      }
-    } finally {
-      setSaving(false)
-    }
-  }, [processedRoutes, CITY_NAME, segmentOffsets, globalStepSize])
-
-  const filteredRoutes = useMemo(() => {
-    if (!processedRoutes) return processedRoutes
-    if (routeStates.size === 0) return processedRoutes
-    const features = processedRoutes.features
-      .map(f => {
-        const segName = (f.properties as any).segmentName as string
-        const state = routeStates.get(segName)
-        if (state === 'hidden') return null
-        if (state === 'magenta') return { ...f, properties: { ...f.properties, color: '#ff00ff' } }
-        return f
-      })
-      .filter(Boolean) as typeof processedRoutes.features
-    return { ...processedRoutes, features }
-  }, [processedRoutes, routeStates])
-
   useEffect(() => {
     if (foundProportion > BEG_THRESHOLD && !hasShownStripeModal) {
       // once we reach a certain threshold, we show the stripe modal
@@ -317,7 +181,7 @@ export default function GamePage({
       if (MAP_FROM_DATA && routes) {
         mapboxMap.addSource('lines', {
           type: 'geojson',
-          data: filteredRoutesRef.current ?? routes,
+          data: routes,
         })
 
         mapboxMap.addLayer({
@@ -547,9 +411,7 @@ export default function GamePage({
       setMap(null)
       mapboxMap.remove()
     }
-  // routes intentionally excluded — updates flow via filteredRoutesRef.current/setData, not map reinit
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [setMap, fc, LINES, MAP_CONFIG, MAP_FROM_DATA])
+  }, [setMap, fc, LINES, MAP_CONFIG, MAP_FROM_DATA, routes])
 
   useEffect(() => {
     if (!map) {
@@ -571,14 +433,6 @@ export default function GamePage({
       map.setFeatureState({ source: 'features', id }, { found: true })
     }
   }, [found, map])
-
-  // Keep the lines source up-to-date when filteredRoutes changes
-  useEffect(() => {
-    filteredRoutesRef.current = filteredRoutes
-    if (!map || !filteredRoutes || !MAP_FROM_DATA) return
-    const src = map.getSource('lines') as mapboxgl.GeoJSONSource | undefined
-    if (src) src.setData(filteredRoutes as any)
-  }, [filteredRoutes, map, MAP_FROM_DATA])
 
   const zoomToFeature = useCallback(
     (id: number) => {
@@ -604,18 +458,7 @@ export default function GamePage({
   )
 
   return (
-    <div className="relative flex h-screen flex-row items-top justify-between">
-      {showSavingOverlay && (
-        <div className="absolute inset-0 z-50 flex items-center justify-center bg-black/40">
-          <div className="rounded-lg bg-white px-6 py-4 shadow-xl text-sm font-medium text-gray-700 flex items-center gap-3">
-            <svg className="animate-spin h-4 w-4 text-gray-500" viewBox="0 0 24 24" fill="none">
-              <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4" />
-              <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8v8H4z" />
-            </svg>
-            {saving ? 'Saving routes…' : 'Loading map…'}
-          </div>
-        </div>
-      )}
+    <div className="flex h-screen flex-row items-top justify-between">
       <div className="relative flex h-screen grow justify-center">
         <div className="absolute left-0 top-0 h-screen w-full" id="map" />
         <div className="absolute top-4 h-12 w-96 max-w-full px-1 lg:top-32">
@@ -644,175 +487,6 @@ export default function GamePage({
             />
           </div>
         </div>
-
-        {MAP_FROM_DATA && processedRoutes && segmentsByLine && segmentInfoMap && (
-          <div className="absolute bottom-4 left-4 z-10 flex flex-col max-w-sm" style={{ maxHeight: 'calc(100vh - 2rem)' }}>
-            <div className="overflow-y-auto rounded-lg bg-white shadow-md text-xs">
-              {/* Global step size control */}
-              <div className="px-3 pt-3 pb-2 border-b border-gray-100">
-                <div className="flex items-center gap-2">
-                  <span className="text-[10px] font-bold text-gray-500 uppercase tracking-wider whitespace-nowrap">Step size</span>
-                  <input
-                    type="range"
-                    min={1}
-                    max={10}
-                    step={0.5}
-                    value={globalStepSize}
-                    onChange={e => setGlobalStepSize(Number(e.target.value))}
-                    className="flex-1 h-1 accent-gray-600"
-                  />
-                  <span className="text-[10px] font-mono text-gray-600 w-6 text-right">{globalStepSize}</span>
-                </div>
-              </div>
-
-              {/* Per-line collapsible sections */}
-              <div className="p-2 flex flex-col gap-1">
-                {[...segmentsByLine.entries()].map(([lineKey, segNames]) => {
-                  const lineConfig = LINES[lineKey]
-                  if (!lineConfig) return null
-                  const isExpanded = expandedLines.has(lineKey)
-                  const color = lineConfig.color
-
-                  return (
-                    <div key={lineKey} className="rounded border border-gray-100 overflow-hidden">
-                      {/* Collapsible line header */}
-                      <button
-                        className="flex items-center gap-1.5 w-full px-2 py-1.5 text-left text-xs font-semibold text-white"
-                        style={{ backgroundColor: color }}
-                        onClick={() => setExpandedLines(prev => {
-                          const next = new Set(prev)
-                          if (next.has(lineKey)) next.delete(lineKey)
-                          else next.add(lineKey)
-                          return next
-                        })}
-                      >
-                        <span className="flex-1">{lineConfig.name}</span>
-                        <span className="text-[10px] opacity-70">{isExpanded ? '▲' : '▼'}</span>
-                      </button>
-
-                      {/* Segment rows */}
-                      {isExpanded && (
-                        <div className="divide-y divide-gray-50">
-                          {segNames.map(segName => {
-                            const isExplicit = segmentOffsets.has(segName)
-                            const interpolated = interpolatedOffsets.get(segName) ?? 0
-                            const displayValue = interpolated
-                            const segInfo = segmentInfoMap.get(segName)
-                            const isOverlap = segInfo?.type === 'overlap'
-
-                            const routeState = routeStates.get(segName)
-
-                            return (
-                              <div key={segName} className="flex items-center gap-1.5 px-2 py-1">
-                                {/* Visibility toggle */}
-                                <button
-                                  className="flex-shrink-0 w-4 h-4 rounded-full border flex items-center justify-center"
-                                  style={
-                                    routeState === 'magenta'
-                                      ? { backgroundColor: '#ff00ff', borderColor: '#ff00ff' }
-                                      : routeState === 'hidden'
-                                      ? { backgroundColor: 'transparent', borderColor: '#d1d5db', borderStyle: 'dashed' }
-                                      : { backgroundColor: color, borderColor: color }
-                                  }
-                                  title={routeState === 'magenta' ? 'Highlighted' : routeState === 'hidden' ? 'Hidden' : 'Visible'}
-                                  onClick={() => setRouteStates(prev => {
-                                    const next = new Map(prev)
-                                    const current = next.get(segName)
-                                    if (!current) next.set(segName, 'magenta')
-                                    else if (current === 'magenta') next.set(segName, 'hidden')
-                                    else next.delete(segName)
-                                    return next
-                                  })}
-                                />
-                                {/* Segment name */}
-                                <span
-                                  className={`font-mono text-[10px] flex-1 truncate ${routeState === 'hidden' ? 'line-through text-gray-300' : isOverlap ? 'text-gray-700 font-semibold' : 'text-gray-400'}`}
-                                  title={segName}
-                                >
-                                  {segName}
-                                </span>
-
-                                {/* Offset controls */}
-                                <div className="flex items-center gap-0.5 flex-shrink-0">
-                                  <button
-                                    className="w-5 h-5 flex items-center justify-center rounded text-gray-600 hover:bg-gray-100 disabled:opacity-30 text-[11px] font-bold"
-                                    disabled={displayValue <= -5}
-                                    onClick={() => setSegmentOffsets(prev => {
-                                      const next = new Map(prev)
-                                      next.set(segName, Math.max(-5, (Math.ceil(displayValue * 2) - 1) / 2))
-                                      return next
-                                    })}
-                                  >−</button>
-                                  <span
-                                    className={`w-8 text-center text-[10px] font-mono tabular-nums ${isExplicit ? 'text-gray-900 font-semibold' : 'text-gray-400'}`}
-                                  >
-                                    {displayValue % 1 === 0 ? displayValue.toFixed(0) : displayValue.toFixed(1)}
-                                  </span>
-                                  <button
-                                    className="w-5 h-5 flex items-center justify-center rounded text-gray-600 hover:bg-gray-100 disabled:opacity-30 text-[11px] font-bold"
-                                    disabled={displayValue >= 5}
-                                    onClick={() => setSegmentOffsets(prev => {
-                                      const next = new Map(prev)
-                                      next.set(segName, Math.min(5, (Math.floor(displayValue * 2) + 1) / 2))
-                                      return next
-                                    })}
-                                  >+</button>
-                                  {/* Reset button — only shown when explicitly set */}
-                                  <button
-                                    className={`w-5 h-5 flex items-center justify-center rounded text-[10px] transition-opacity ${isExplicit ? 'text-gray-400 hover:text-red-400 hover:bg-red-50' : 'opacity-0 pointer-events-none'}`}
-                                    onClick={() => setSegmentOffsets(prev => {
-                                      const next = new Map(prev)
-                                      next.delete(segName)
-                                      return next
-                                    })}
-                                    title="Reset to interpolated"
-                                  >×</button>
-                                </div>
-                              </div>
-                            )
-                          })}
-                        </div>
-                      )}
-                    </div>
-                  )
-                })}
-              </div>
-
-              {/* Actions */}
-              <div className="px-3 py-2 border-t border-gray-100 flex flex-col gap-1.5">
-                {latestSavedSettings && !savedSettingsMatchCurrent && (
-                  <button
-                    className="w-full rounded px-3 py-1.5 text-xs font-semibold border transition-colors"
-                    style={{ backgroundColor: '#f9fafb', borderColor: '#d1d5db', color: '#374151' }}
-                    onClick={handleLoadSavedSettings}
-                  >
-                    Load saved settings
-                  </button>
-                )}
-                {hasChanges && (
-                  <button
-                    className="w-full rounded px-3 py-1.5 text-xs font-semibold border transition-colors"
-                    style={{ backgroundColor: '#fff5f5', borderColor: '#fca5a5', color: '#b91c1c' }}
-                    onClick={handleReset}
-                  >
-                    Reset all changes
-                  </button>
-                )}
-                <button
-                  className="w-full rounded px-3 py-1.5 text-xs font-semibold text-white transition-colors disabled:opacity-40"
-                  style={{
-                    backgroundColor: saveSuccess ? '#16a34a' : hasChanges ? '#374151' : '#9ca3af',
-                    cursor: hasChanges && !saving ? 'pointer' : 'not-allowed',
-                  }}
-                  disabled={!hasChanges || saving}
-                  onClick={handleSaveRoutes}
-                >
-                  {saving ? 'Saving…' : saveSuccess ? 'Saved!' : 'Save processed routes'}
-                </button>
-              </div>
-            </div>
-          </div>
-        )}
       </div>
       <div className="z-10 hidden h-full overflow-y-auto bg-zinc-50 p-6 shadow-lg lg:block lg:w-96 xl:w-[32rem]">
         <FoundSummary
